split bindable into root class
