+++ Assets/Scripts/ALM/Util/UIToolkitExtend/DataBinderCS.cs
@@ -1,0 +1,10 @@
+namespace ALM.Util.UIToolkitExtend
+{
+    public class DataBinderCS
+    {
+        public DataBinderCS()
+        {
+            
+        }
+    }
+}
