+++ Assets/Scripts/ALM/Util/UIToolkitExtend/Mono/Bindable/Bindable.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 2c39f6fc08dae3d4fa6ea253f0c150e8
